// frontend/src/api/models/aTATimelineEntryDTO.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * API
 * Clean Architecture API with JWT Authentication
 * OpenAPI spec version: v1
 */
import type { ATAStatusDTO } from './aTAStatusDTO';

export interface ATATimelineEntryDTO {
  id?: string;
  status?: ATAStatusDTO;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  changedBy?: string | null;
  /** @nullable */
  changedByName?: string | null;
  timestamp?: string;
  submissionRound?: number;
}